+++ cargo-workspaces/src/rename.rs
@@ -3,14 +3,14 @@ use crate::utils::{
     GroupName, WorkspaceConfig,
 };
 use cargo_metadata::Metadata;
-use clap::{ArgSettings, Parser};
-use glob::{Pattern, PatternError};
+use clap::Parser;
+use globset::{Error as GlobsetError, Glob};
 use std::{collections::BTreeMap as Map, fs};
 
 /// Rename crates in the project
 #[derive(Debug, Parser)]
 pub struct Rename {
-    /// Rename private creates too
+    /// Rename private crates too
     #[clap(short, long)]
     pub all: bool,
 
@@ -18,18 +18,19 @@ pub struct Rename {
     #[clap(long, value_name = "pattern")]
     pub ignore: Option<String>,
 
+    /// Rename only a specific crate
+    #[clap(short, long, value_name = "crate", conflicts_with_all = &["all", "ignore"])]
+    pub from: Option<String>,
+
     /// The value that should be used as new name (should contain `%n`)
-    #[clap(
-        validator = validate_value_containing_name,
-        setting(ArgSettings::ForbidEmptyValues)
-    )]
+    #[clap(forbid_empty_values(true))]
     pub to: String,
 
     /// Comma separated list of crate groups to rename
     #[clap(
         long,
         multiple_occurrences = true,
-        use_delimiter = true,
+        use_value_delimiter = true,
         number_of_values = 1
     )]
     pub groups: Vec<GroupName>,
@@ -39,29 +40,47 @@ impl Rename {
     pub fn run(self, metadata: Metadata) -> Result<(), Error> {
         let config: WorkspaceConfig = read_config(&metadata.workspace_metadata)?;
 
-        let workspace_groups = get_group_packages(&metadata, &config, self.all)?;
+        let workspace_groups =
+            get_group_packages(&metadata, &config, self.all || self.from.is_some())?;
 
         let ignore = self
             .ignore
             .clone()
-            .map(|x| Pattern::new(&x))
-            .map_or::<Result<_, PatternError>, _>(Ok(None), |x| Ok(x.ok()))?;
+            .map(|x| Glob::new(&x))
+            .map_or::<Result<_, GlobsetError>, _>(Ok(None), |x| Ok(x.ok()))?;
 
         let mut rename_map = Map::new();
 
-        for ((group_name, _), pkg) in workspace_groups.into_iter() {
-            if let Some(pattern) = &ignore {
-                if pattern.matches(&pkg.name) {
+        if let Some(from) = self.from {
+            if workspace_groups
+                .into_iter()
+                .map(|(_, p)| p.name)
+                .collect::<Vec<_>>()
+                .contains(&&from)
+            {
+                rename_map.insert(from, self.to.clone());
+            } else {
+                return Err(Error::PackageNotFound { id: from });
+            }
+        } else {
+            // Validate the `to` value
+            validate_value_containing_name(&self.to)
+                .map_err(|_| Error::MustContainPercentN("<TO>".into()))?;
+
+            for ((group_name, _), pkg) in workspace_groups.into_iter() {
+                if let Some(pattern) = &ignore {
+                    if pattern.compile_matcher().is_match(&pkg.name) {
+                        continue;
+                    }
+                }
+                if !(self.groups.is_empty() || self.groups.contains(&group_name)) {
                     continue;
                 }
-            }
-            if !(self.groups.is_empty() || self.groups.contains(&group_name)) {
-                continue;
-            }
 
-            let new_name = self.to.replace("%n", &pkg.name);
+                let new_name = self.to.replace("%n", &pkg.name);
 
-            rename_map.insert(pkg.name, new_name);
+                rename_map.insert(pkg.name, new_name);
+            }
         }
 
         for pkg in &metadata.packages {